requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.11.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.100.0
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
JPype1==1.5.0
konlpy==0.6.0
lxml==5.2.2
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.25.1
openai==1.35.3
packaging==24.1
pip-autoremove==0.10.0
pip-check-reqs==2.5.3
psycopg2-binary==2.9.6
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
python-dotenv==1.0.0
requests==2.32.3
scikit-learn==1.3.0
scipy==1.13.1
sniffio==1.3.1
SQLAlchemy==2.0.19
starlette==0.27.0
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.22.0
Werkzeug==3.0.3
yarl==1.9.4
